requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
filelock==3.3.2
idna==3.3
lxml==4.6.3
MechanicalSoup==1.1.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
six==1.16.0
soupsieve==2.2.1
tldextract==3.1.2
tqdm==4.62.3
urllib3==1.26.7
